Support for Arch Linux added. May need further testing on Debian for stability

+++ setup.sh
@@ -24,29 +24,36 @@ function config() {
 }
 
 function install() {
-    echo "Installing necessary packages..."
-    read -p 'Perform apt-get update? (y/n): ' update
-    if [ "${update^^}" = "Y" ]
-    then
-      $SUDO apt-get update
+    OS=$(hostnamectl | grep -i "operating system")
+    if [[ $OS =~ "Arch Linux" ]]; then
+        $SUDO pacman -Qi python > /dev/null ||  (echo "Updating Python..." && $SUDO pacman -S python)
+        $SUDO pacman -Qi python-pip > /dev/null ||  $SUDO pacman -S python-pip
+        $SUDO pacman -Qi chromium > /dev/null || $SUDO pacman -S chromium
+    else
+        echo "Installing necessary packages..."
+        read -p 'Perform apt-get update? (y/n): ' update
+        if [ "${update^^}" = "Y" ]
+        then
+            $SUDO apt-get update
+        fi
+
+        $SUDO apt -y install chromium-chromedriver || \
+        $SUDO apt -y install chromium-driver || \
+        $SUDO apt -y install chromedriver
+
+        PYV=`python3 -c "import sys;t='{v[0]}{v[1]}'.format(v=list(sys.version_info[:2]));sys.stdout.write(t)";`
+        if [[ "$PYV" -lt "36" ]] || ! hash python3;
+        then
+            echo "This script requires Python version 3.6 or higher. Attempting to install..."
+            $SUDO apt-get -y install python3
+        fi
+
+        # Debian9 package 'python-selenium' does not work with chromedriver,
+        # Install from pip, which is newer
+        $SUDO apt -y install chromium-browser # Update Chromium Browser or script won't work.
+        $SUDO apt -y install $PYTHON-pip
+
     fi
-
-    $SUDO apt -y install chromium-chromedriver || \
-      $SUDO apt -y install chromium-driver || \
-      $SUDO apt -y install chromedriver
-
-
-    PYV=`python3 -c "import sys;t='{v[0]}{v[1]}'.format(v=list(sys.version_info[:2]));sys.stdout.write(t)";`
-    if [[ "$PYV" -lt "36" ]] || ! hash python3;
-    then
-      echo "This script requires Python version 3.6 or higher. Attempting to install..."
-      $SUDO apt-get -y install python3
-    fi
-
-    # Debian9 package 'python-selenium' does not work with chromedriver,
-    # Install from pip, which is newer
-    $SUDO apt -y install chromium-browser # Update Chromium Browser or script won't work.
-    $SUDO apt -y install $PYTHON-pip
     $SUDO $PYTHON -m pip install selenium
 }
 
@@ -67,8 +74,7 @@ function deploy() {
     $SUDO chown $USER $INSTDIR/noip-renew-skd.sh
     $SUDO chmod 700 $INSTEXE
     noip
-    $SUDO sed -i '/noip-renew/d' /etc/crontab
-    echo "$CRONJOB" | $SUDO tee -a /etc/crontab
+    ($SUDO crontab -u $USER -l; echo "$CRONJOB") | $SUDO crontab -u $USER -
     $SUDO sed -i 's/USER=/USER='$USER'/1' $INSTDIR/noip-renew-skd.sh
     echo "Installation Complete."
     echo "To change noip.com account details, please run setup.sh again."
